Refactor inter-page methods & menu item handling
Promises are now used to simplify headless assembly/compilation and 
error handling.

The HTML id of the menu LI is used as the url for that page. The text 
content of the LI is then used as the title. Also new onload event 
watcher ensures that the page title is always correct.