[emu] Re-design call instruction
Call now only saves the current PC to the new frame instead of PC and
SP.

This greatly simplifies the microcode and microinstructions needed to
implement the instruction.

Additionally, by requiring stack pointers to be even, the PC setting
logic from the goto instruction can be re-used, providing further
simplification to the control unit.