[cmp] Check for missing return statements
If a function is not type 'none' then throw an error if it does not
contain a return statement. No control flow analysis is used so the
return statement is not guaranteed to always be executed. Eg. it could
be placed in an if statement which may not execute it.