[emu,asm,cmp] Simplify call instruction
Remove the "store PC of next instruction to new SP" part of the call
instruction.

This makes it the same as the same as the return instruction. The
assembler treats it as an alias.

This brings the following benefits:
- All instructions take 1 cycle in the execute stage
- The PC->data bus microinstruction is no longer needed

The compiler simply inserts another write instruction to store the
return PC address into callee stack frame.

Very small perf impact in most cases.